requirements.txt
faiss-cpu
sentence-transformers
streamlit
openai
python-dotenv